Refactor example metadata parsing utilities(#20204)
