+++ datafusion-examples/src/utils/mod.rs
@@ -17,6 +17,6 @@
 
 mod csv_to_parquet;
 pub mod datasets;
-pub mod examples_docs;
+pub mod example_metadata;
 
 pub use csv_to_parquet::write_csv_to_parquet;
